Refactor: useFetchUserIntfo 훅 분리
